### 2019-07-07 #1552 @labarba
👋 @heilerich — your paper is not compiling. Please add the required bibliography field, according to the [documentation](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).
