### 2019-07-07 #1552 @labarba
@cMadan — Would you add another paper to your editing load? This one seems closest to your areas of domain knowledge.
