### 2019-07-07 #1552 @brainstorm
@satra, would you help me out reviewing this submission for JOSS? :)
