### 2019-07-08 #1552 @brainstorm
Argh, true, forgot about that... perhaps @chrisgorgo can be our second reviewer for JOSS? 👋
