### 2019-07-16 #1552 @brainstorm
@rougier Would you like to assist reviewing this paper?
