### 2019-07-16 #1552 @rougier
I'm already late with another JOSS review but I can try do to this one by end of July if it is ok.
