### 2019-07-22 #1552 @brainstorm
Thanks @rougier! It's fine for me unless @heilerich is in a rush to get this review done?
