### 2019-08-05 #1626 @whedon
[REVIEW]: neurone_loader: A Python module for loading EEG data recorded with Bittium NeurOne
**Submitting author:** @heilerich (<a href="http://orcid.org/0000-0002-9068-2679">Felix Heilmeyer</a>)
**Repository:** <a href="https://github.com/heilerich/neurone_loader" target ="_blank">https://github.com/heilerich/neurone_loader</a>
**Version:** v1.1.0
