### 2019-08-13 #1626 @brainstorm
@rougier @satra Let me know if you have any kind of doubts/issues while conducting this review ;)
