### 2019-08-20 #1626 @brainstorm
Thanks @satra! I noticed that you didn't check the versioning box... is it because it does not follow [semver](https://semver.org)? (has a `v` as a prefix). If so, @heilerich, could you stick to the [semver](https://semver.org) spec on your releases?
