### 2019-08-20 #1626 @brainstorm
@rougier no worries, see you next week!

@satra Do you reckon that the "statement of need" is not clear/complete?
