### 2019-08-20 #1626 @satra
@brainstorm

- the latest release has not been tagged or pushed to github. pypi says 1.1.0, but github still has 1.0.1
- statement of need is in the pdf of the paper but not in the documentation itself. target audience is in neither and could be made more explicit.
