### 2019-08-20 #1626 @brainstorm
Thanks Satra, good points. @heilerich, could you please address the points raised please?
