### 2019-08-25 #1626 @rougier
I'm now reviewing the paper and software. I managed to install the software but I get trouble testing it because I can't get download the data (I opened an issue).  Who is the second author (I cannot find any commit by him)? I also noticed the discrepancy between the GitHub release version (1.0.1) and the release given here (1.1.0).
