### 2019-08-25 #1626 @rougier
Problem solved.  But tests are failing: 41 failed, 12 passed, 21 error in 3.13 seconds
Test dependencies: zodbpickle, braindecode, mne

Example error:
```
FileNotFoundError: [Errno 2] No such file or directory: '/Users/rougier/tmp/neurone_loader/test/test_data/ses1_1_events.npy'
```
test search in `test_data` while it seems to be  in data are in `test/data/NeuroneTestdata`. I ran the get_test_data.sh from inside the test drectory, maybe this is the reason.

Also, it would be good to warn the user (in the README) that the test data is very large (> 2.75 Go).
