### 2019-08-25 57b593b
Clarify testing procedure in readme, update changelog
