+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.1.0
+
+* New channel dropping feature
+* Bugfixes for the export to python-mne
+
+# 1.0.1
+
+* Improved test stability
+* Unified logging for all modules
+
 # 1.0.0
 
 * Initial release
