### 2019-08-25 2e04909
purpose statement in documentation intro
