### 2019-08-25 #1626 @heilerich
Thanks for your feedback everyone!

**Release Tag**
@brainstorm @rougier  I have tagged the release on github and also removed the v prefix to comply with semver.

**Statement of Need and Audience**
@satra I added a short statement regarding the purpose of the library to the documentation introduction (see the [review branch](https://github.com/heilerich/neurone_loader/tree/review)). I could add more of the paper contents to the docs but I would prefer to keep it short. Would you be satisfied with the changes I made?
Regarding the target audience I must admit I a little unsure as to how make it more explicit. It is simply anyone working with python and NeurOne equipment. That could include but is not limited to neuroscientists, other researchers, doctors, clinical staff ... Do you think it is necessary to explicitly state that in the documentation?

**Testing**
@rougier The test data should indeed be in the `test/test_data` folder. I added a note to the README that the scripts should be run from the main directory. I also added a note regarding the size of the test data (see the [review branch](https://github.com/heilerich/neurone_loader/tree/review)).
Were you able to run the tests in the meantime?

**Authors**
@rougier The second author is my supervisor. He provided the equipment and the data used in the development. He also provided some (unpublished) reference implementations for loading and converting the data. Some data converted with these implementations are used in the tests as a reference to validate the loading results.
