### 2019-08-26 #1626 @rougier
Tests are now working. I get 6 failed, 68 passed in 695.95 seconds. Difficult to say where it failed because the output is quite verbose. Last failed test was:
```
___________________________________ TestAgainstBBCI.test_length ____________________________________

self = <test.test_mne_export.TestAgainstBBCI testMethod=test_length>

    def test_length(self):
>       self.assertEqual(self.bbci_cnt.last_samp, self.raw_cnt.last_samp)
E       AssertionError: 33462 != 31198
```
Apart from that I think the submission is now ok for me.
