### 2019-09-04 #1626 @heilerich
Thanks for testing. I ran the tests again on my machines and everything passed. A [CI Build](https://travis-ci.org/heilerich/neurone_loader/builds/580714188) of the most recent commit is also still passing. I suspect there is some kind of incompatibility regarding the environment. In any case, it seems you might have encountered an actual bug and I would really appreciate your help in resolving this.
Maybe you could open another issue in the [repository](https://github.com/heilerich/neurone_loader/) and provide me with some details including your OS, python version, installed packages (e.g. output of `pip freeze`) and the complete output of the tests (you might consider placing that in a separate gist for the sake of readability). Thanks again and have a nice day 👋🏻
