### 2019-09-05 #1626 @brainstorm
@satra Does it tick the remaining boxes for you according to https://github.com/openjournals/joss-reviews/issues/1626#issuecomment-524649967?
