### 2019-09-21 #1626 @brainstorm
Looks all good to ship @arfon.
