### 2019-10-01 bfeedb5
readme test data size fix
