+++ README.rst
@@ -95,7 +95,7 @@ Testing
 To run the tests you must first get the test data and then you can run the test with the following commands.
 Please run them in the repository directory, not in the test subdirectory.
 
-To get the test data (~2.7GB) you need to install wget_. Then you can 
+To get the test data (~2.8GB) you need to install wget_. Then you can 
 download the data by running
 
 .. code:: bash
