+++ .github/workflows/pythonpackage.yml
@@ -1,0 +1,32 @@
+name: Python package
+
+on:
+  release
+  types: [published]
+
+evn:
+  PYTHON_VERSION: 3.7
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v1
+    - name: Set up Python ${{ env.PYTHON_VERSION }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ env.PYTHON_VERSION }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Build package
+      run: |
+        setup.py bdist_wheel --universal
+    - name: Publish package
+      uses: pypa/gh-action-pypi-publish@master
+      with:
+        user: __token__
+        password: ${{ secrets.pypi_secret }}
