+++ .github/workflows/pythonpackage.yml
@@ -4,9 +4,6 @@ on:
   release:
     types: [published]
 
-env:
-  PYTHON_VERSION: 3.7
-
 jobs:
   build:
 
@@ -14,10 +11,10 @@ jobs:
 
     steps:
     - uses: actions/checkout@v1
-    - name: Set up Python ${{ env.PYTHON_VERSION }}
+    - name: Set up Python 3.7
       uses: actions/setup-python@v1
       with:
-        python-version: ${{ env.PYTHON_VERSION }}
+        python-version: 3.7
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
