### 2019-10-01 49cc2d2
missing command in github workflow
