+++ .github/workflows/pythonpackage.yml
@@ -21,7 +21,7 @@ jobs:
         pip install -r requirements.txt
     - name: Build package
       run: |
-        setup.py bdist_wheel --universal
+        python setup.py bdist_wheel --universal
     - name: Publish package
       uses: pypa/gh-action-pypi-publish@master
       with:
