+++ .github/workflows/pythonpackage.yml
@@ -19,6 +19,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
+        pip install wheel
     - name: Build package
       run: |
         python setup.py bdist_wheel --universal
