### 2019-10-01 #1626 @heilerich
@arfon Okay, released, tagged, deployed, archived ;)
I hope everything works now. The DOI of the zenodo archive is `10.5281/zenodo.3468534`
